backend/requirements.txt
fastapi[all]==0.115.0
sqlmodel==0.0.21
alembic==1.13.3
psycopg2-binary>=2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
apscheduler==3.10.4
slowapi==0.1.9
